requirement.txt
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.6
django-allauth==0.36.0
django-cors-headers==2.2.0
django-rest-auth==0.9.3
django-rest-registration==0.3.12
django-widget-tweaks==1.4.2
djangorestframework==3.8.2
freeze==1.0.10
idna==2.7
oauthlib==2.1.0
Pillow==5.1.0
python3-openid==3.1.0
pytz==2018.4
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
urllib3==1.23
